requirements.txt
openai
streamlit
python-dotenv
pyttsx3
langchain
PyPDF2
python-docx
python-pptx
Pillow
pytesseract
requests
beautifulsoup4
lxml
readability-lxml
langchain-community
gtts
pypdf
googletrans
textwrap
asyncio
inspect